Add and use NSCoder in NSOpenGLPixelFormat
This makes it possible to load and initialize a NSOpenGLView from a NIB
with no modifications from OSX. NSOpenGLPixelFormat directly implements
initWithCoder so that it may marshall the pixel attributes prior to
initWithAttributes, but encodeWithCoder is simply passed to the back-end
implementation (pending further implementation).

Tested:
Built and successfully ran a Cocoa application using a custom NSOpenGLView
in a NIB designed with Xcode 4.5's Interface Builder.